Cascade delete everywhere

+++ controller/db_review.py
@@ -20,37 +20,37 @@ def create_review(db: Session, request: ReviewBase):
 
 def get_review(db: Session, id: int):
     review = db.query(DbReview).filter(DbReview.id == id).first()
-    # Handle errors
     if not review:
-        raise HTTPException(status_code= status.HTTP_404_NOT_FOUND,
-                             detail = f'review with id {id} is not exist!')
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                             detail=f'Review with id {id} does not exist!')
     return review
 
 
 def get_all_reviews(db: Session, user_id: int):
     review = db.query(DbReview).filter(DbReview.user_id == user_id).all()
-    # Handle errors
+    if not review:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                             detail=f'Review for user_id {user_id} does not exist!')
     return review
 
 
 def get_all_reviews_left(db: Session, creator_id: int):
     review = db.query(DbReview).filter(DbReview.creator_id == creator_id).all()
-    # Handle errors
     if not review:
-        raise HTTPException(status_code= status.HTTP_404_NOT_FOUND,
-                             detail = f'review with id {id} is not exist!')
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                             detail =f'Review with creator_id {creator_id} does not exist!')
     return review
 
 
 def update_review(db: Session, id: int, request: ReviewBase):
     review = db.query(DbReview).filter(DbReview.id == id)
+    if not review:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                             detail=f'Review with id {id} does not exist!')
     review.update({
         DbReview.mark: request.mark,
         DbReview.text_description: request.text_description,
-        DbReview.user_id: request.user_id,
-        DbReview.creator_id: request.creator_id,
         DbReview.created_at: func.now()
-
     })
     db.commit()
     return 'ok'
@@ -58,6 +58,9 @@ def update_review(db: Session, id: int, request: ReviewBase):
 
 def delete_review(db: Session, id: int):
     review = db.query(DbReview).filter(DbReview.id == id).first()
+    if not review:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                             detail=f'Review with id {id} does not exist!')
     db.delete(review)
     db.commit()
     return "ok"
